fix(TUP-21433):Studio autologin not work.
