+++ main/plugins/org.talend.repository/src/main/java/org/talend/repository/ui/login/LoginHelper.java
@@ -474,6 +474,8 @@ public class LoginHelper {
         if (connBean == null || project == null || project.getLabel() == null) {
             return false;
         }
+        LoginFetchLicenseHelper loginFetchLicenseHelper = new LoginFetchLicenseHelper();
+        loginFetchLicenseHelper.fetchLicenseIfNeeded(project);
         setCurrentSelectedConnBean(connBean);
         try {
             if (!project.getEmfProject().isLocal() && factory.isLocalConnectionProvider()) {
